+++ src/AStarSearch.cpp
@@ -39,7 +39,7 @@ AStarSearch::AStarSearch(Map* pMap)
 void AStarSearch::Search(Map* pMap, UnitBase* pUnit, Coord start, Coord destination) {
 #if 1
     std::ranges::fill(mapData, TileData{});
-    std::ranges::fill(depthCheckCount, 0);
+    std::ranges::fill(depthCheckCount, static_cast<short>(0));
 #else
     // Clobber PODs the evil way...
     memset(&mapData[0], 0, sizeof(mapData[0]) * mapData.size());
