+++ src/GUI/RadioButton.cpp
@@ -105,7 +105,7 @@ void RadioButton::draw(Point position) {
         return;
     }
 
-    tex.draw(dune::globals::renderer.get(), position.x, position.y);
+    tex.draw(dune::globals::renderer.get(), static_cast<float>(position.x), static_cast<float>(position.y));
 }
 
 void RadioButton::resize(uint32_t width, uint32_t height) {
